Split go package
