+++ mage.go
@@ -1,3 +1,4 @@
+//go:build ignore
 // +build ignore
 
 package main
