+++ pkg/gav3/grafana.go
@@ -1,16 +1,16 @@
-package main
+package gav3
 
 import (
 	"fmt"
-	"sort"
-	"strconv"
-	"strings"
-	"time"
-
 	"github.com/grafana/grafana-plugin-sdk-go/backend/log"
 	"github.com/grafana/grafana-plugin-sdk-go/data"
 	"github.com/jinzhu/copier"
 	reporting "google.golang.org/api/analyticsreporting/v4"
+	"sort"
+	"strconv"
+	"strings"
+	"time"
+	"util"
 )
 
 func transformReportToDataFrameByDimensions(columns []*ColumnDefinition, rows []*reporting.ReportRow, refId string, dimensions string) (*data.Frame, error) {
@@ -173,10 +173,6 @@ func transformReportsResponseToDataFrames(reportsResponse *reporting.GetReportsR
 	return &frames, nil
 }
 
-func padRightSide(str string, item string, count int) string {
-	return str + strings.Repeat(item, count)
-}
-
 // timeConverter handles sheets TIME column types.
 var timeConverter = data.FieldConverter{
 	OutputFieldType: data.FieldTypeNullableTime,
@@ -254,14 +250,14 @@ func copyRow(row *reporting.ReportRow) *reporting.ReportRow {
 
 func copyRowAndInit(row *reporting.ReportRow) *reporting.ReportRow {
 	copyRow := copyRow(row)
-	copyRow.Metrics[0].Values = FillArray(make([]string, len(row.Metrics[0].Values)), "0")
+	copyRow.Metrics[0].Values = util.FillArray(make([]string, len(row.Metrics[0].Values)), "0")
 	return copyRow
 }
 
 func parseRow(row *reporting.ReportRow, timezone *time.Location) (*reporting.ReportRow, *time.Time) {
 	timeDimension := row.Dimensions[0]
 	otherDimensions := row.Dimensions[1:]
-	parsedTime, err := ParseAndTimezoneTime(timeDimension, timezone)
+	parsedTime, err := util.ParseAndTimezoneTime(timeDimension, timezone)
 	if err != nil {
 		log.DefaultLogger.Error("parsedTime err", "err", err.Error())
 	}
@@ -277,20 +273,20 @@ func getTimeFunction(timeDimension string) (func(time.Time) time.Time, func(time
 	var add, sub func(time.Time) time.Time
 	switch timeDimension {
 	case timeDimensions[0]:
-		add = AddOneMinute
-		sub = SubOneMinute
+		add = util.AddOneMinute
+		sub = util.SubOneMinute
 		break
 	case timeDimensions[1]:
-		add = AddOneHour
-		sub = SubOneHour
+		add = util.AddOneHour
+		sub = util.SubOneHour
 		break
 	case timeDimensions[2]:
-		add = AddOneDay
-		sub = SubOneDay
+		add = util.AddOneDay
+		sub = util.SubOneDay
 		break
 	default:
-		add = AddOneHour
-		sub = SubOneHour
+		add = util.AddOneHour
+		sub = util.SubOneHour
 	}
 	return add, sub
 }
