+++ pkg/util/util.go
@@ -3,6 +3,8 @@ package main
 import (
 	"github.com/grafana/grafana-plugin-sdk-go/backend/log"
 	"time"
+	"strings"
+
 )
 
 func Elapsed(what string) func() {
@@ -10,6 +12,10 @@ func Elapsed(what string) func() {
 	return func() {
 		log.DefaultLogger.Debug("Elapsed", what, time.Since(start).String())
 	}
+}
+
+func padRightSide(str string, item string, count int) string {
+	return str + strings.Repeat(item, count)
 }
 
 func ParseAndTimezoneTime(sTime string, timezone *time.Location) (*time.Time, error) {
